rename: 예외 패키지 구조 변경
