+++ src/main/java/org/promgrammers/springbootbasic/domain/voucher/repository/impl/FileVoucherRepository.java
@@ -2,8 +2,8 @@ package org.promgrammers.springbootbasic.domain.voucher.repository.impl;
 
 import org.promgrammers.springbootbasic.domain.voucher.model.Voucher;
 import org.promgrammers.springbootbasic.domain.voucher.repository.VoucherRepository;
-import org.promgrammers.springbootbasic.exception.InvalidFilePathException;
-import org.promgrammers.springbootbasic.exception.VoucherFileWriteException;
+import org.promgrammers.springbootbasic.exception.repository.InvalidFilePathException;
+import org.promgrammers.springbootbasic.exception.repository.VoucherFileWriteException;
 import org.promgrammers.springbootbasic.util.FileConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
