+++ src/main/java/org/promgrammers/springbootbasic/exception/repository/DoesNotExistException.java
@@ -1,0 +1,8 @@
+package org.promgrammers.springbootbasic.exception.repository;
+
+public class DoesNotExistException extends RuntimeException {
+
+    public DoesNotExistException(String message) {
+        super(message);
+    }
+}
